\name{PercolationIndicator-methods}
\alias{print.PercolationIndicator}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Methods for PercolationIndicator objects
}
\description{
Print method prints the PercolationIndicator output.
}
\usage{
\method{print}{PercolationIndicator}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
output of \code{\link{PercolationIndicator}}
}
  \item{\dots}{
Currently not used.
}
}
\author{
Claudia van Borkulo
}